Removed flower length from model. Removed subvestita from data.
Flower length is highly correlated with seed mass and flowers per head.

Subvestita grows well outside the range of the other species included
in the analysis.

Set WinBUGS to run in this directory so that I can pull out DIC results
from log file manually (for component-wise DIC comparisons as well as
overall).